+++ import_export/resources.py
@@ -1027,13 +1027,24 @@ class Resource(metaclass=DeclarativeMetaclass):
     def get_export_fields(self):
         return [self.fields[f] for f in self.get_export_order()]
 
-    def export_resource(self, instance):
-        return [
-            self.export_field(field, instance) for field in self.get_export_fields()
-        ]
+    def export_resource(self, instance, fields=None):
+        export_fields = self.get_export_fields()
 
-    def get_export_headers(self):
+        if isinstance(fields, list) and fields:
+            return [
+                self.export_field(field, instance)
+                for field in export_fields
+                if field.column_name in fields
+            ]
+
+        return [self.export_field(field, instance) for field in export_fields]
+
+    def get_export_headers(self, fields=None):
         headers = [force_str(field.column_name) for field in self.get_export_fields()]
+
+        if isinstance(fields, list) and fields:
+            return [f for f in headers if f in fields]
+
         return headers
 
     def get_user_visible_headers(self):
@@ -1075,11 +1086,12 @@ class Resource(metaclass=DeclarativeMetaclass):
         if queryset is None:
             queryset = self.get_queryset()
         queryset = self.filter_export(queryset, **kwargs)
-        headers = self.get_export_headers()
+        export_fields = kwargs.get("export_fields", None)
+        headers = self.get_export_headers(fields=export_fields)
         dataset = tablib.Dataset(headers=headers)
 
         for obj in self.iter_queryset(queryset):
-            dataset.append(self.export_resource(obj))
+            dataset.append(self.export_resource(obj, fields=export_fields))
 
         self.after_export(queryset, dataset, **kwargs)